feat: SteamParsingService 데이터 추출 로직 구현 feat: SteamCollectorFacade 통합 수집 흐름 완성
test: SteamParsingService 데이터 파싱 단위 테스트 추가
test: SteamCollectorFacade API 연동 통합 테스트 추가

+++ src/main/java/com/example/steam_tracker/game/entity/Game.java
@@ -34,7 +34,7 @@ public class Game extends BaseEntity {
     private int discountPercent;
 
     @Column(nullable = false)
-    private int trendScore7d;
+    private double trendScore7d;
 
     @Column(nullable = false)
     boolean isActive = true;
